+++ changelog.md
@@ -4,6 +4,12 @@
 - Recalculate totalInvoiced for ClientSlice after update operation
 - add invoice view/print page
 
+# 2022 03 18
+- created ui for filtering and sorting invoices
+
+# 2022 03 19
+- Add support in invoice store slice to filter and sort invoices.
+
 # next task
 - add context menu to invoice table
 - Refactor/organize based on [slide for Week 5 Office Hour 1](https://docs.google.com/presentation/d/1pLZ8Zvpyr_myotpvyVYzY5awpUFUO78DE5CnBnVNm7c/edit#slide=id.g117828bad90_0_9)
@@ -13,3 +19,6 @@
 - add email, phone
 - taxes to invoices?
 - soft deletion?
+- sorted filters? currently api will sort always in the same order. 
+  it would be nice to give the user the control of choosing which sorting
+  comes first.